+++ resources/views/bootstrap-components/media/image.blade.php
@@ -1,13 +1,13 @@
 <div {{ htmlAttributes($containerId ? ['id' => $containerId] : null) }}
-    {{ classTag('image-container', $containerClass) }}
+    {{ classTag('image-container', $containerClasses) }}
     {{ htmlAttributes($containerHtmlAttributes) }}>
     <a {{ htmlAttributes($linkId ? ['id' => $linkId] : null) }}
         {{ htmlAttributes($linkUrl ? 'href="'.$linkUrl.'"' : null) }}
-        {{ classTag('image-link', $linkClass) }}
+        {{ classTag('image-link', $linkClasses) }}
         {{ htmlAttributes($linkHtmlAttributes) }}
         {{ htmlAttributes($alt ? 'title="'.$alt.'"' : null) }}>
         <img {{ htmlAttributes($componentId ? ['id' => $componentId] : null) }}
-            {{ classTag('image-component', $componentClass) }}
+            {{ classTag('image-component', $componentClasses) }}
             {{ htmlAttributes($width ? 'width="'.$width.'"' : null) }}
             {{ htmlAttributes($height ? 'height="'.$height.'"' : null) }}
             {{ htmlAttributes($componentHtmlAttributes) }}