+++ tests/Unit/Button/BackTest.php
@@ -212,40 +212,40 @@ class BackTest extends BootstrapComponentsTestCase
         $this->assertStringContainsString('<a id="' . $customComponentId . '"', $html);
     }
 
-    public function testConfigContainerClass()
+    public function testConfigContainerClasses()
     {
-        $configContainerCLass = 'test-config-class-container';
-        config()->set('bootstrap-components.button.back.class.container', [$configContainerCLass]);
+        $configContainerClasses = 'test-config-class-container';
+        config()->set('bootstrap-components.button.back.class.container', [$configContainerClasses]);
         $html = bsBack()->toHtml();
-        $this->assertStringContainsString('class="button-container ' . $configContainerCLass . '">', $html);
+        $this->assertStringContainsString('class="button-container ' . $configContainerClasses . '">', $html);
     }
 
-    public function testSetContainerClass()
+    public function testSetContainerClasses()
     {
-        $configContainerCLass = 'test-config-class-container';
-        $customContainerCLass = 'test-custom-class-container';
-        config()->set('bootstrap-components.input.class.container', [$configContainerCLass]);
-        $html = bsBack()->containerClass([$customContainerCLass])->toHtml();
-        $this->assertStringContainsString('class="button-container ' . $customContainerCLass . '">', $html);
-        $this->assertStringNotContainsString('class="button-container ' . $configContainerCLass . '">', $html);
+        $configContainerClasses = 'test-config-class-container';
+        $customContainerClasses = 'test-custom-class-container';
+        config()->set('bootstrap-components.input.class.container', [$configContainerClasses]);
+        $html = bsBack()->containerClasses([$customContainerClasses])->toHtml();
+        $this->assertStringContainsString('class="button-container ' . $customContainerClasses . '">', $html);
+        $this->assertStringNotContainsString('class="button-container ' . $configContainerClasses . '">', $html);
     }
 
     public function testConfigComponentClass()
     {
-        $configComponentCLass = 'test-config-class-component';
-        config()->set('bootstrap-components.button.back.class.component', [$configComponentCLass]);
+        $configComponentClasses = 'test-config-class-component';
+        config()->set('bootstrap-components.button.back.class.component', [$configComponentClasses]);
         $html = bsBack()->toHtml();
-        $this->assertStringContainsString('class="button-component ' . $configComponentCLass . '"', $html);
+        $this->assertStringContainsString('class="button-component ' . $configComponentClasses . '"', $html);
     }
 
     public function testSetComponentClass()
     {
-        $configComponentCLass = 'test-config-class-component';
-        $customComponentCLass = 'test-custom-class-component';
-        config()->set('bootstrap-components.button.back.class.component', [$customComponentCLass]);
-        $html = bsBack()->componentClass([$customComponentCLass])->toHtml();
-        $this->assertStringContainsString('class="button-component ' . $customComponentCLass . '"', $html);
-        $this->assertStringNotContainsString('class="button-component ' . $configComponentCLass . '"', $html);
+        $configComponentClasses = 'test-config-class-component';
+        $customComponentClasses = 'test-custom-class-component';
+        config()->set('bootstrap-components.button.back.class.component', [$customComponentClasses]);
+        $html = bsBack()->componentClasses([$customComponentClasses])->toHtml();
+        $this->assertStringContainsString('class="button-component ' . $customComponentClasses . '"', $html);
+        $this->assertStringNotContainsString('class="button-component ' . $configComponentClasses . '"', $html);
     }
 
     public function testConfigContainerHtmlAttributes()