+++ tests/Unit/Media/AudioTest.php
@@ -67,40 +67,40 @@ class AudioTest extends BootstrapComponentsTestCase
         $this->assertStringContainsString('<audio id="' . $customComponentId . '"', $html);
     }
 
-    public function testConfigContainerClass()
+    public function testConfigContainerClasses()
     {
-        $configContainerCLass = 'test-config-class-container';
-        config()->set('bootstrap-components.media.audio.class.container', [$configContainerCLass]);
+        $configContainerClasses = 'test-config-class-container';
+        config()->set('bootstrap-components.media.audio.class.container', [$configContainerClasses]);
         $html = audio()->toHtml();
-        $this->assertStringContainsString('class="audio-container ' . $configContainerCLass . '"', $html);
+        $this->assertStringContainsString('class="audio-container ' . $configContainerClasses . '"', $html);
     }
 
-    public function testSetContainerClass()
+    public function testSetContainerClasses()
     {
-        $configContainerCLass = 'test-config-class-container';
-        $customContainerCLass = 'test-custom-class-container';
-        config()->set('bootstrap-components.media.audio.class.container', [$configContainerCLass]);
-        $html = audio()->containerClass([$customContainerCLass])->toHtml();
-        $this->assertStringContainsString('class="audio-container ' . $customContainerCLass . '"', $html);
-        $this->assertStringNotContainsString('class="audio-container ' . $configContainerCLass . '"', $html);
+        $configContainerClasses = 'test-config-class-container';
+        $customContainerClasses = 'test-custom-class-container';
+        config()->set('bootstrap-components.media.audio.class.container', [$configContainerClasses]);
+        $html = audio()->containerClasses([$customContainerClasses])->toHtml();
+        $this->assertStringContainsString('class="audio-container ' . $customContainerClasses . '"', $html);
+        $this->assertStringNotContainsString('class="audio-container ' . $configContainerClasses . '"', $html);
     }
 
     public function testConfigComponentClass()
     {
-        $configComponentCLass = 'test-config-class-component';
-        config()->set('bootstrap-components.media.audio.class.component', [$configComponentCLass]);
+        $configComponentClasses = 'test-config-class-component';
+        config()->set('bootstrap-components.media.audio.class.component', [$configComponentClasses]);
         $html = audio()->toHtml();
-        $this->assertStringContainsString('class="audio-component ' . $configComponentCLass . '"', $html);
+        $this->assertStringContainsString('class="audio-component ' . $configComponentClasses . '"', $html);
     }
 
     public function testSetComponentClass()
     {
-        $configComponentCLass = 'test-config-class-component';
-        $customComponentCLass = 'test-custom-class-component';
-        config()->set('bootstrap-components.media.audio.class.component', [$customComponentCLass]);
-        $html = audio()->componentClass([$customComponentCLass])->toHtml();
-        $this->assertStringContainsString('class="audio-component ' . $customComponentCLass . '"', $html);
-        $this->assertStringNotContainsString('class="audio-component ' . $configComponentCLass . '"', $html);
+        $configComponentClasses = 'test-config-class-component';
+        $customComponentClasses = 'test-custom-class-component';
+        config()->set('bootstrap-components.media.audio.class.component', [$customComponentClasses]);
+        $html = audio()->componentClasses([$customComponentClasses])->toHtml();
+        $this->assertStringContainsString('class="audio-component ' . $customComponentClasses . '"', $html);
+        $this->assertStringNotContainsString('class="audio-component ' . $configComponentClasses . '"', $html);
     }
 
     public function testConfigContainerHtmlAttributes()